Add better memory access functions to MMU API.
These functions allow stable access to RAM ranges with and without
translation.

Unlike ReadFromHardware() and WriteFromHardware() and functions that
call those, these here are not sensitive to MSR.DR and can therefore be
called regardless of whether the emulation is inside user or interrupt
handling code, which make them ideal for cheating and similar purposes.